fix(gpu): signed EIP-2200 refund counter, EIP-3529 cap at dispatcher
Before this fix the kernel refund_counter was unsigned (ulong / unsigned
long long). The EIP-2200 "subsequent modification with original != 0
and current == 0" branch executes `rc -= GAS_SSTORE_REFUND`; when rc was
0 (or smaller than 4800) the unsigned subtraction underflowed to ~2^64,
the dispatcher subtracted that from gas_used, and the result blew past
the gas_limit. Real-world Solidity contracts hit this on slot
clear-then-set patterns within a single tx.

Type changes (signed accumulator across all four backends):
  Metal kernel:  refund_counter → long, sstore_gas_eip2200 ref → thread long&,
                 TxOutput.gas_refund → long.
  CUDA kernel:   refund_counter → long long, sstore_gas_eip2200 ref → long long&,
                 TxOutput.gas_refund → long long.
  CPU interp.:   InterpreterResult.gas_refund (new) = int64_t, refund_counter
                 local = int64_t, sstore_gas_eip2200 takes int64_t&. The CPU
                 interpreter previously did NOT track refunds at all — it now
                 mirrors the kernel-side computation byte-for-byte and is the
                 parity reference for Metal/CUDA.
  Host structs:  Metal/CUDA TxOutput.gas_refund and TxResult.gas_refund are
                 int64_t. Wire layout (sizeof) is unchanged; only the
                 signedness of the same 8 bytes changed.

EIP-3529 cap (gas_used / 5) is applied centrally in
evm::gpu::apply_eip3529_cap (gpu_dispatch.cpp). Each kernel emits the raw
signed refund per tx; the dispatcher floors at 0, caps at gas_used / 5,
and subtracts. Three call sites updated: execute_kernel_cpu, run_metal,
run_cuda. Kernels stay backend-agnostic; cap policy lives in one place.

Tests (test/unittests/refund_test.cpp, registered as refund-test):
  1. CreateSlotNoRefund — SSTORE 0→non-zero earns no refund.
  2. ClearSlotEarnsRefundUpToCap — clear-after-set in same tx triggers
     the EIP-2200 restoration branch (orig=0); cap holds final refund
     to gas_used / 5.
  3. NoUnderflowOnRepeatedClear — the original bug reproducer:
     0→0x42→0→0x99→0 used to overflow into 2^64 territory, now stays
     under 1M gas.
  4. Eip3529CapApplied — explicit assertion that uncapped refund (19900)
     is reduced to gas_used/5.
  5. ParityWithEvmoneOnClearSlot — both backends earn a positive refund
     and produce gas_used well under gas_limit. Exact-number parity is
     out of scope here because the kernel doesn't model EIP-2929
     cold/warm SLOAD pricing yet (separate fix).

All 5 tests pass; existing dispatch-test (15 tests), stack-depth-test
(3), metal-host-test (6), evm-parity-test (133 vectors) all green.

Wire ABI: TxOutput sizeof unchanged. Go bridge surface (`gas_used` post-
cap) is byte-identical to the prior contract. ABI version stays at 4.

+++ lib/evm/gpu/kernel/evm_kernel.metal
@@ -350,8 +350,12 @@ struct TxInput {
     uint code_offset; uint code_size; uint calldata_offset; uint calldata_size;
     ulong gas_limit; uint256 caller; uint256 address; uint256 value;
 };
+// gas_refund is signed: SSTORE may transiently subtract refund credit
+// (clear-then-set-then-clear within a tx). The kernel emits the raw
+// signed value; the host dispatcher floors at 0 and applies the EIP-3529
+// cap of gas_used/5.
 struct TxOutput {
-    uint status; ulong gas_used; ulong gas_refund; uint output_size;
+    uint status; ulong gas_used; long gas_refund; uint output_size;
 };
 struct StorageEntry { uint256 key; uint256 value; };
 
@@ -415,11 +419,15 @@ static inline bool original_value_lookup(thread OriginalEntry* o,uint n,uint256 
 static inline void original_value_record(thread OriginalEntry* o,thread uint& n,uint256 s,uint256 v){
     for(uint i=0;i<n;++i)if(o[i].valid&&u256_eq(o[i].key,s))return;
     if(n<MAX_STORAGE_PER_TX){o[n].key=s;o[n].value=v;o[n].valid=true;n++;}}
-static inline ulong sstore_gas_eip2200(uint256 orig,uint256 cur,uint256 nv,thread ulong& rc){
+// EIP-2200 net gas SSTORE accounting. `rc` accumulates the *signed* refund
+// delta — clear-then-set-then-clear within a single tx produces a transient
+// negative value. The dispatcher applies EIP-3529 (max refund = gas_used/5)
+// and floors at 0 after execution.
+static inline ulong sstore_gas_eip2200(uint256 orig,uint256 cur,uint256 nv,thread long& rc){
     if(u256_eq(nv,cur))return GAS_SSTORE_NOOP;
-    if(u256_eq(orig,cur)){if(u256_iszero(orig))return GAS_SSTORE_SET;if(u256_iszero(nv))rc+=GAS_SSTORE_REFUND;return GAS_SSTORE_RESET;}
-    if(!u256_iszero(orig)){if(u256_iszero(cur))rc-=GAS_SSTORE_REFUND;else if(u256_iszero(nv))rc+=GAS_SSTORE_REFUND;}
-    if(u256_eq(nv,orig)){if(u256_iszero(orig))rc+=GAS_SSTORE_SET-GAS_SSTORE_NOOP;else rc+=GAS_SSTORE_RESET-GAS_SSTORE_NOOP;}
+    if(u256_eq(orig,cur)){if(u256_iszero(orig))return GAS_SSTORE_SET;if(u256_iszero(nv))rc+=long(GAS_SSTORE_REFUND);return GAS_SSTORE_RESET;}
+    if(!u256_iszero(orig)){if(u256_iszero(cur))rc-=long(GAS_SSTORE_REFUND);else if(u256_iszero(nv))rc+=long(GAS_SSTORE_REFUND);}
+    if(u256_eq(nv,orig)){if(u256_iszero(orig))rc+=long(GAS_SSTORE_SET-GAS_SSTORE_NOOP);else rc+=long(GAS_SSTORE_RESET-GAS_SSTORE_NOOP);}
     return GAS_SSTORE_NOOP;
 }
 
@@ -504,7 +512,7 @@ kernel void evm_execute(
     uint256 stack[1024];
     uint sp = 0;
     ulong gas = inp.gas_limit;
-    ulong refund_counter = 0;
+    long  refund_counter = 0;
     uint pc = 0;
     uint mem_size = 0;
     ulong gas_start = gas;
